temporal/md/requirements.txt
# Core dependencies
requests==2.31.0
click==8.1.7
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Progress tracking
tqdm==4.66.1

# Data Science & ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Holidays
holidays==0.38

# Model Persistence
joblib==1.3.2

# Jupyter
# jupyter==1.0.0
# jupyterlab==4.0.9
# notebook==7.0.6
ipykernel>=6.29.0
